Move Sentry stuff to own file for readability
